+++ libbaghash/options.h
@@ -5,9 +5,10 @@
 
 int options_validate (struct baghash_options *opt);
 
-size_t options_n_blocks (const struct baghash_options *opts);
+uint64_t options_n_blocks (const struct baghash_options *opts);
 
-size_t options_block_size (const struct baghash_options *opts);
+uint16_t options_block_size (const struct baghash_options *opts);
 
-unsigned int options_n_neighbors (const struct baghash_options *opts);
+uint8_t options_n_neighbors (const struct baghash_options *opts);
+
 #endif
